Implemented naf agent, working trading but having
issues with change of customer scale. first small then large leads to NaN, first large then
small leads to loss of profitability. Overall, the agent seems to be
capable of generating profits if it is given a producer but it does not
achieve the same amount of profit as the average price of the market
would have achieved. I.e. the broker is not capable of achieving average
market performance.